Fix a potential integer overflow.
GCC seems to handle values above 255, but let's better play it safe.

+++ crawl-ref/source/fontwrapper-ft.cc
@@ -292,7 +292,7 @@ ucs_t FTFontWrapper::map_unicode(ucs_t uchar, bool update)
                             pixels[idx] = orig;
                             pixels[idx + 1] = orig;
                             pixels[idx + 2] = orig;
-                            pixels[idx + 3] = std::min(orig + edge, 255);
+                            pixels[idx + 3] = std::min((int)orig + edge, 255);
                         }
                     }
             }
